계좌 관련기능 구성중

+++ src/main/java/kopo/aisw/hc/account/dao/AccountDAO.java
@@ -7,9 +7,9 @@ import kopo.aisw.hc.transaction.vo.TransactionVO;
 
 public interface AccountDAO {
 	//계좌 개설- 계좌번호 생성-자동 (수동은 추후 업데이트 예정)
-	public double createAccNum(int productNum);
+	public long createAccNum(int productNum);
 	//계좌 개설- 생성된 계좌번호 중복체크
-	public boolean accDoubleCheck(double accNum);
+	public boolean accDoubleCheck(long accNum);
 	//계좌 개설-등록
 	public boolean openAnAccount(AccountVO account);
 
@@ -17,9 +17,9 @@ public interface AccountDAO {
 	public List<AccountVO> getAccList(int customerId);
 	//추후 accNum 대신 accId로 조회할 예정
 	//계좌 조회- 이름 체크
-	public String getOwnerName(double accNum);
+	public String getOwnerName(long accNum);
 	//계좌 조회- 잔액 조회
-	public double getBalance(double accNum);
+	public long getBalance(long accNum);
 	//입금
 	public boolean deposit(TransactionVO transaction);
 	//출금
